+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 Breaking Change.
 
 ### Added
+- mongo_class.RepSetColl.coll_dst:  Query of document using distinct command.
 - mongo_class.RepSetColl.coll_find:  Query of document using find command.
 - mongo_class.RepSetColl.\_db_auth:  Database authentication, private function for connect method.